% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{base_enr_url}
\alias{base_enr_url}
\title{STRING API URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
base_enr_url
}
\description{
Base STRING API URL for constructing enrichment tsv requests
}
\keyword{datasets}
