% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{base_string_url}
\alias{base_string_url}
\title{STRING API URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
base_string_url
}
\value{
\code{https://string-db.org/api}
}
\description{
Base STRING DB API URL for constructing requests
}
\examples{
pullString:::base_string_url
}
\keyword{datasets}
