% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_png_response.R
\name{.create_api_png_resp}
\alias{.create_api_png_resp}
\title{Function for creating other functions that call STRING API endpoints}
\usage{
.create_api_png_resp(endpoint)
}
\arguments{
\item{endpoint}{STRING API endpoint}
}
\description{
This is a higher-order function that is not a part of user interface. It is
internally used to create other functions that call STRING API endpoints and
collect \code{image} response
}
