% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{.create_api_tsv_resp}
\alias{.create_api_tsv_resp}
\title{Function for creating other functions that call STRING API endpoints}
\usage{
.create_api_tsv_resp(endpoint)
}
\arguments{
\item{endpoint}{STRING API endpoint}
}
\value{
\code{function} to perform STRING API request
}
\description{
This is a higher-order function that is not a part of the user interface. It
is internally used to create other functions that call STRING API endpoints
and collect \code{tsv} response
}
