% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_enrichment_tab}
\alias{endpoint_enrichment_tab}
\title{STRING endpoint for enrichment table}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_enrichment_tab
}
\description{
STRING endpoint for enrichment table
}
\keyword{datasets}
