% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_get_best_homology}
\alias{endpoint_get_best_homology}
\title{STRING endpoint for best protein homology}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_get_best_homology
}
\description{
STRING endpoint for best protein homology
}
\keyword{datasets}
