% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_get_enrichment_plot}
\alias{endpoint_get_enrichment_plot}
\title{STRING endpoint for values/ranks enrichment
STRING endpoint for enrichment plot}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_get_enrichment_plot
}
\description{
STRING endpoint for values/ranks enrichment
STRING endpoint for enrichment plot
}
\keyword{datasets}
