% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_get_functional_annotation}
\alias{endpoint_get_functional_annotation}
\title{STRING endpoint for functional annotation}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_get_functional_annotation
}
\description{
STRING endpoint for functional annotation
}
\keyword{datasets}
