% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_get_highres_png_network}
\alias{endpoint_get_highres_png_network}
\title{STRING endpoint for high resolution \code{.png} network}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_get_highres_png_network
}
\description{
STRING endpoint for high resolution \code{.png} network
}
\keyword{datasets}
