% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_get_homology_scores}
\alias{endpoint_get_homology_scores}
\title{STRING endpoint for protein similarity scores (homology)}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_get_homology_scores
}
\description{
STRING endpoint for protein similarity scores (homology)
}
\keyword{datasets}
