% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_get_interaction_enrichment}
\alias{endpoint_get_interaction_enrichment}
\title{STRING endpoint for interaction enrichment}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_get_interaction_enrichment
}
\description{
STRING endpoint for interaction enrichment
}
\keyword{datasets}
