% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_get_link}
\alias{endpoint_get_link}
\title{STRING endpoint for generating STRING URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_get_link
}
\description{
STRING endpoint for generating STRING URL
}
\keyword{datasets}
