% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{endpoint_ids}
\alias{endpoint_ids}
\title{STRING endpoint for mapping identifiers}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_ids
}
\description{
STRING endpoint for mapping identifiers
}
\keyword{datasets}
