% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_interact_partners}
\alias{endpoint_interact_partners}
\title{STRING endpoint for interaction partners}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_interact_partners
}
\description{
STRING endpoint for interaction partners
}
\keyword{datasets}
