% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_net_interactions}
\alias{endpoint_net_interactions}
\title{STRING endpoint for network interactions}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_net_interactions
}
\description{
STRING endpoint for network interactions
}
\keyword{datasets}
