% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{endpoint_version}
\alias{endpoint_version}
\title{STRING endpoint for current stable version}
\format{
An object of class \code{character} of length 1.
}
\usage{
endpoint_version
}
\description{
STRING endpoint for current stable version
}
\keyword{datasets}
