% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{get_best_homology}
\alias{get_best_homology}
\title{Retrieving best similarity hits between species}
\usage{
get_best_homology(ids, species = 9606, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{...}{other params passed to STRING API request, options:}
}
\description{
STRING internally uses the Smith–Waterman bit scores as a proxy for protein
homology. The original scores are computed by \code{SIMILARITY MATRIX OF PROTEINS
(SIMAP)} project. Using this API you can retrieve these scores between the
proteins in a selected species.
}
\section{Other arguments}{

\itemize{
  \item \code{species_b} - NCBI taxon identifier, default is \code{9606}
  (human)
  \item \code{caller_identity} - optional, your identity for \code{STRING DB}
  developers. If left blank, defaults to \code{pullString} and current
  package version
}
}

