% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enr.R
\name{get_enr}
\alias{get_enr}
\title{Get STRING enrichment data.frame for list of genes}
\usage{
get_enr(gene, species = 9606, ...)
}
\arguments{
\item{gene}{character vector of genes to plot on a STRING net}

\item{species}{STRING taxon identifier, default homo sapiens \code{9606}}

\item{...}{other arguments passed to \code{\link[httr:POST]{httr::POST()}}}
}
\value{
enrichment results \code{data frame}
}
\description{
Get STRING enrichment data.frame for list of genes
}
