% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{get_enrichment}
\alias{get_enrichment}
\title{Get functional enrichment of your identifiers set}
\usage{
get_enrichment(ids, species = 9606, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{...}{other params passed to STRING API request, options:}
}
\description{
The STRING enrichment function allows you to retrieve functional enrichment
for any set of input proteins. It will tell you which of your input proteins
have an enriched term and the term's description.
}
\section{Other arguments}{

\itemize{
  \item \code{caller_identity} - optional, your identity for \code{STRING DB}
  developers. If left blank, defaults to \code{pullString} and current
  package version
}
}

\section{Details}{
 STRING maps several databases onto its proteins, this includes:
  \code{Gene Ontology, KEGG pathways, UniProt Keywords, PubMed publications,
  Pfam domains, InterPro domains, SMART domains}. \code{NOTE:} As with other
  STRING APIs, we automatically expand the network by 10 proteins when a
  query includes only one protein. Keep in mind that this network expansion
  will likely reduce the FDR more than expected by chance and does not
  indicate an enrichment of the original single-protein set but rather of its
  immediate interaction neighborhood. When querying with a set of two or more
  proteins, no additional proteins are added to the input, and the FDR
  accurately reflects the probability of observing the enrichment by random
  chance.
}

\section{Value}{
 \code{tibble} with terms and categories, \code{p-values} and
  \code{FDR}
}

