% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_png_response.R
\name{get_enrichment_plot}
\alias{get_enrichment_plot}
\title{Get enrichment plot}
\usage{
get_enrichment_plot(ids, species = 9606, png_filename, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{png_filename}{path to save \code{png} file with network}

\item{...}{other params passed to \code{STRING API} request, options:}
}
\description{
This function enables the visualization of enrichment analysis, providing a
way to explore results across three key dimensions:
\itemize{
\item Enrichment signal (\code{x-axis})
\item False Discovery Rate (\code{FDR}) represented by the dot color
\item Protein count in the network indicated by the dot size
}
}
\section{Other arguments}{

\itemize{
  \item \code{category} -term category (e.g., KEGG, WikiPathways, etc. See
  the table below for all category keys. Default is \code{Process})
  \item \code{group_by_similarity} - threshold for visually grouping related
  terms on the plot, ranging from \code{0.1 to 1}, in steps of \code{0.1}
  (e.g. \code{0.8)}, with no grouping applied by default
  \item \code{color_palette} - color palette to represent FDR values
  (\code{mint_blue, lime_emerald, green_blue, peach_purple, straw_navy,
  yellow_pink}, default is \code{mint_blue})
  \item \code{number_of_term_shown} - maximum number of terms displayed on
  the plot (default is \code{10})
  \item \code{x_axis} - specifies the order of the terms and the variable on
  the x-axis (\code{signal, strength, FDR, gene_count}, default is
  \code{signal})
}
}

\section{Details}{

Category keys:
\itemize{
  \item \code{Process} - Biological Process (Gene Ontology)
  \item \code{Function} - 	Molecular Function (Gene Ontology)
  \item \code{Component} - Cellular Component (Gene Ontology)
  \item \code{Keyword} - Annotated Keywords (UniProt)
  \item \code{KEGG} - KEGG Pathways
  \item \code{RCTM} - Reactome Pathways
  \item \code{HPO} - Human Phenotype (Monarch)
  \item \code{MPO} - he Mammalian Phenotype Ontology (Monarch)
  \item \code{DPO} - Drosophila Phenotype (Monarch)
  \item \code{WPO} - C. elegans Phenotype Ontology (Monarch)
  \item \code{ZPO} - Zebrafish Phenotype Ontology (Monarch)
  \item \code{FYPO} - Fission Yeast Phenotype Ontology (Monarch)
  \item \code{Pfam} - Protein Domains (Pfam)
  \item \code{SMART} - Protein Domains (SMART)
  \item \code{InterPro} - Protein Domains and Features (InterPro)
  \item \code{PMID} - Reference Publications (PubMed)
  \item \code{NetworkNeighborAL} - Local Network Cluster (STRING)
  \item \code{COMPARTMENTS} - Subcellular Localization (COMPARTMENTS)
  \item \code{TISSUES} - Tissue Expression (TISSUES)
  \item \code{DISEASES} - Disease-gene Associations (DISEASES)
  \item \code{WikiPathways} - WikiPathways
}
}

\examples{
\dontrun{
get_enrichment_plot(
  ids = c("POSTN", "IFNA1", "SERPINB2"),
  png_filename = "string_enrichment_fig.png"
)
}
}
