% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{get_functional_annotation}
\alias{get_functional_annotation}
\title{Retrieve functional annotation}
\usage{
get_functional_annotation(ids, species = 9606, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{...}{other params passed to STRING API request, options:}
}
\description{
You can retrieve all these annotations (and not only enriched subset) for
your proteins
}
\section{Other arguments}{

\itemize{
  \item \code{allow_pubmed} - \code{0} or \code{1} to print also the PubMed
  annotations in addition to other categories, default is \code{0}
  \item \code{only_pubmed} - \code{0} or \code{1} to print only PubMed
  annotations,default is \code{0}
  \item \code{caller_identity} - optional, your identity for \code{STRING DB}
  developers. If left blank, defaults to \code{pullString} and current
  package version
}
}

\section{Details}{
 STRING maps several databases onto its proteins, this
  includes: \code{Gene Ontology, KEGG pathways, UniProt Keywords, PubMed
  publications, Pfam domains, InterPro domains, SMART domains}. \code{NOTE:}
  Due to the potential large size of the \code{PubMed (Reference
  Publications)} assignments, they won't be sent by default, but you can turn
  them back on by specifying the \code{allow_pubmed=1} parameter.
}

\section{Value}{

 \code{tibble} with terms and categories, number and ratio of the proteins in
 your input list with the term assigned
}

