% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_png_response.R
\name{get_highres_png_network}
\alias{get_highres_png_network}
\title{Retrieve higher resolution \code{png} network}
\usage{
get_highres_png_network(ids, species = 9606, png_filename, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{png_filename}{path to save \code{png} file with network}

\item{...}{other params passed to \code{STRING API} request, options:}
}
\description{
Retrieve an image of a STRING network of a neighborhood surrounding one or
more proteins or ask STRING to show only the network of interactions between
your input proteins. Both the network flavors (confidence and evidence) and
network types (functional and physical) are accessible through the API.
}
\section{Other arguments}{

\itemize{
  \item \code{add_color_nodes} - adds color nodes based on scores to the
  \item \code{add_white_nodes} - adds white nodes based on scores to the
  input proteins
  \item \code{required_score} - threshold of significance to include an
  interaction, a number between 0 and 1000 (default depends on the network)
  number between 0 and 1000 (default depends on the network)
  \item \code{network_type} - network type: \code{functional} (default),
  \code{physical}
  \item \code{network_flavor} - the style of edges in the network:
  \code{evidence} (default), \code{confidence, actions}
  \item \code{hide_node_labels} - hides all protein names from the picture
  \code{(0 or 1)} default is \code{0}
  \item \code{hide_disconnected_nodes} - hides all
  proteins that are not connected to any other protein in your network
  \code{(0 or 1)}, default is \code{0} labels in the network image \code{(0
  or 1)}, default is \code{0}
  \item \code{block_structure_pics_in_bubbles} - disables structure pictures
  inside the bubble \code{(0 or 1)} (default is \code{0})
  \item \code{flat_node_design} - disable 3D bubble design \code{(0
  or 1)}, default is \code{0}
  \item \code{center_node_labels} - center
  protein names on nodes \code{(0 or 1)}, default is \code{0}
  \item
  \code{custom_label_font_size} - change font size of the protein names
  \code{(from 5 to 50)}, default is \code{12}
  \item \code{caller_identity} -
  optional, your identity for \code{STRING DB} developers. If left blank,
  defaults to \code{pullString} and current package version
}
}

\section{Details}{
 If you query the API with one protein the
  \code{add_white_nodes} parameter is automatically set to \code{10}, so you
  can see the interaction neighborhood of your query protein. However,
  similarly to the STRING webpage, whenever you query the API with more than
  one protein we show only the interactions between your input proteins. You
  can, of course, always extend the interaction neighborhood by setting
  \code{add_color} or \code{white_nodes} parameter to the desired value.
}

\examples{
\dontrun{
get_png_network(
  ids = c("POSTN", "IFNA1", "SERPINB2", "CXCL10"),
  png_filename = "./string_net.png"
)
}
}
