% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{get_homology}
\alias{get_homology}
\title{Retrieving similarity scores of the protein set}
\usage{
get_homology(ids, species = 9606, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{...}{other params passed to \code{STRING API} request, options:}
}
\description{
STRING internally uses the Smith–Waterman bit scores as a proxy for protein
homology. The original scores are computed by \code{SIMILARITY MATRIX OF PROTEINS
(SIMAP)} project. Using this API you can retrieve these scores between the
proteins in a selected species.
}
\section{Other arguments}{

\itemize{
  \item \code{caller_identity} - optional, your identity for \code{STRING DB}
  developers. If left blank, defaults to \code{pullString} and current
  package version
}
}

\section{Details}{
 Scores are symmetric, therefore to make the transfer a bit
  faster we will send only half of the similarity matrix (\code{A -> B}, but
  not symmetric \code{B -> A}) and the self-hits. The \code{bit score}
  cut-off below which we do not store or report homology is 50.
}

\section{Value}{
 \code{tibble} with score between STRING ID A and B
}

\examples{
\dontrun{
get_homology(c("POSTN",  "SERPINB2"))
}
}
