% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{get_interact_partners}
\alias{get_interact_partners}
\title{Get all the STRING interaction partners of your proteins}
\usage{
get_interact_partners(ids, species = 9606, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{...}{other params passed to STRING API request, options:}
}
\description{
Diffrently from the \code{\link[pullString]{get_net_interactions}}
interactions between the set of input proteins and between their closest
interaction neighborhood (if \code{add_nodes} parameters is specified),
\code{get_interact_partners} fucntion provides the interactions between your
set of proteins and all the other STRING proteins. As STRING network usually
has a lot of low scoring interactions, you may want to limit the number of
retrieved interaction per protein using "limit" parameter (of course the high
scoring interactions will come first).
}
\section{Other arguments}{

\itemize{
  \item \code{limit} - limits the number of interaction partners retrieved
  per protein (most confident interactions come first)
  \item \code{required_score} - threshold of significance to include a
  interaction, a number between 0 and 1000 (default depends on the network)
  \item \code{network_type} - options: \code{functional} (default), or \code{physical}
  \item \code{caller_identity} - optional, your identity for \code{STRING DB}
  developers. If left blank, defaults to \code{pullString} and current
  package version
}
}

\section{Value}{
 \code{tibble} with STRING preferred name for protein A and B,
 taxon id and scores:
\itemize{
  \item \code{score} - combined score
  \item \code{nscore} - gene neighborhood score
  \item \code{fscore} - gene fusion score
  \item \code{pscore} - phylogenetic profile score
  \item \code{ascore} - coexpression score
  \item \code{escore} - experimental score
  \item \code{dscore} - database score
  \item \code{tscore} - textmining score
}
}

\examples{
\dontrun{
get_interact_partners(c("POSTN",  "SERPINB2"))
}
}
