% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{get_net_interactions}
\alias{get_net_interactions}
\title{Retrieves the network interactions for your input protein(s)}
\usage{
get_net_interactions(ids, species = 9606, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{...}{other params passed to \code{STRING API} request, options:}
}
\description{
The network API function also allows you to retrieve your STRING interaction
network for one or multiple proteins. It will tell you the combined score and
all the channel specific scores for the set of proteins. You can also extend
the network neighborhood by setting \code{add_nodes}, which will add, to your
network, new interaction partners in order of their confidence.
}
\section{Other arguments}{

\itemize{
  \item \code{required_score} - threshold of significance to include an
  interaction, a number between 0 and 1000 (default depends on the network)
  \item \code{network_type} - options: \code{functional} (default), or
  \code{physical} come first)
  \item \code{add_nodes} - a number of proteins with to the network based on
  their confidence score
  \item \code{show_query_node_labels} - when
  available use submitted names in the preferredName column. Takes values
  \code{0} or \code{1}, default is \code{0}
  \item \code{caller_identity} -
  optional, your identity for \code{STRING DB} developers. If left blank,
  defaults to \code{pullString} and current package version
}
}

\section{Value}{
 \code{tibble} with STRING preferred name for protein A and B,
 taxon id and scores:
\itemize{
  \item \code{score} - combined score
  \item \code{nscore} - gene neighborhood score
  \item \code{fscore} - gene fusion score
  \item \code{pscore} - phylogenetic profile score
  \item \code{ascore} - coexpression score
  \item \code{escore} - experimental score
  \item \code{dscore} - database score
  \item \code{tscore} - textmining score
}
}

\examples{
\dontrun{
get_net_interactions(
  c("ENSG00000133110", # POSTN
    "ENSP00000401645", # SERPINB2
    "CXCL10"
  )
)
}
}
