% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{get_ppi_enrichment}
\alias{get_ppi_enrichment}
\title{Retrieve protein-protein interaction enrichment}
\usage{
get_ppi_enrichment(ids, species = 9606, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{...}{other params passed to STRING API request, options:}
}
\description{
Even in the absence of annotated proteins (e.g. in novel genomes) STRING can
tell you if your subset of proteins is functionally related, that is, if it is
enriched in interactions in comparison to the background proteome-wide
interaction distribution.
}
\section{Other arguments}{

\itemize{
  \item \code{required_score} - threshold of significance to include a
  interaction, a number between 0 and 1000 (default depends on the network)
  in addition to other categories, default is \code{0}
  \item \code{only_pubmed} - \code{0} or \code{1} to print only PubMed
  annotations, default is \code{0}
  \item \code{caller_identity} - optional, your identity for \code{STRING DB}
  developers. If left blank, defaults to \code{pullString} and current
  package version
}
}

