% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_string.R
\name{get_string_identifiers}
\alias{get_string_identifiers}
\title{Maps common protein names, synonyms and UniProt identifiers into STRING
identifiers}
\usage{
get_string_identifiers(ids, species = 9606, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{...}{other params passed to API}
}
\value{
\code{tibble} with STRING preferred name, annotation and taxon
  details
}
\description{
Maps common protein names, synonyms and UniProt identifiers into STRING
identifiers
}
\examples{
get_string_identifiers(
  c("ENSG00000133110", # POSTN
    "ENSP00000401645", # SERPINB2
    "CXCL10"
  )
)
}
