% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_png.R
\name{get_string_png}
\alias{get_string_png}
\title{Get STRING .png for list of genes}
\usage{
get_string_png(gene, species = 9606, png_filename, label_font_size = 16, ...)
}
\arguments{
\item{gene}{character vector of genes to plot on a STRING net}

\item{species}{STRING taxon identifier, default homo sapiens \code{9606}}

\item{png_filename}{path to write png}

\item{label_font_size}{font size on save png, default \code{16}}

\item{...}{other arguments passed to \code{\link[httr:POST]{httr::POST()}}}
}
\value{
saved .png to \code{png_filename}
}
\description{
Get STRING .png for list of genes
}
