% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{get_string_version}
\alias{get_string_version}
\title{Get the current STRING version and its stable address}
\usage{
get_string_version()
}
\value{
\code{tibble} with STRING version and current stable website URL
  address
}
\description{
Get the current STRING version and its stable address
}
\examples{
\dontrun{
get_string_version()
}
}
