% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_string_net.R
\name{string_request_js}
\alias{string_request_js}
\title{Define STRING network parameters}
\usage{
string_request_js(
  ids,
  species = 9606,
  add_color_nodes = 3,
  add_white_nodes = 5,
  required_score = 400,
  network_flavor = "confidence",
  network_type = "functional",
  hide_node_labels = 0,
  hide_disconnected_nodes = 0,
  show_query_node_labels = 0,
  block_structure_pics_in_bubbles = 1,
  caller_identity = usr_agent
)
}
\arguments{
\item{ids}{s}

\item{species}{s}

\item{add_color_nodes}{s}

\item{add_white_nodes}{h}

\item{required_score}{h}

\item{network_flavor}{g}

\item{network_type}{g}

\item{hide_node_labels}{f}

\item{hide_disconnected_nodes}{d}

\item{show_query_node_labels}{d}

\item{block_structure_pics_in_bubbles}{d}

\item{caller_identity}{d}
}
\description{
Define STRING network parameters
}
