% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\docType{data}
\name{usr_agent}
\alias{usr_agent}
\title{Set user agent - package name and version}
\format{
An object of class \code{character} of length 1.
}
\usage{
usr_agent
}
\description{
Set user agent - package name and version
}
\examples{
pullString:::usr_agent
}
\keyword{datasets}
