% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv_response.R
\name{get_string_identifiers}
\alias{get_string_identifiers}
\title{Maps common protein names, synonyms and UniProt identifiers into STRING
identifiers}
\usage{
get_string_identifiers(ids, species = 9606, ...)
}
\arguments{
\item{ids}{character vector of identifiers, can be protein names, synonyms,
ensembl id, ensemble protein id, HGNC symbol, UniProt or mixed}

\item{species}{NCBI/STRING taxon, default \code{9606} (human)}

\item{...}{other params passed to \code{STRING API} request, options:}
}
\description{
Maps common protein names, synonyms and UniProt identifiers into STRING
identifiers
}
\section{Other arguments}{

\itemize{
  \item \code{echo_query} - insert column with your input identifier, takes
  values \code{0} or \code{1}, default is \code{0}
  \item \code{limit} - limits the number of matches per query identifier
  (best matches come first)
  \item \code{caller_identity} - optional, your identity for \code{STRING DB}
  developers. If left blank, defaults to \code{pullString} and current
  package version
}
}

\section{Value}{
 \code{tibble} with STRING preferred name, annotation and
  taxon details
}

\examples{
\dontrun{
get_string_identifiers(
  c("ENSG00000133110", # POSTN
    "ENSP00000401645", # SERPINB2
    "CXCL10"
  )
)
}
}
